requirements.txt
loguru==0.7.2
matplotlib==3.9.2
numpy==2.1.2
opencv-python==4.10.0.84
Pillow==10.4.0
ruamel.yaml==0.18.10
scikit-learn==1.5.2
timm==1.0.9
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
ttach==0.0.3
